// src/proto/tokyo2006/cirque/cirque.ts
export interface GetStateRequest {
}

export interface GetStateResponse {
  state: CirqueState | undefined;
}

export interface SetStateRequest {
  state: CirqueState | undefined;
  persist: boolean;
}

export interface SetStateResponse {
  state: CirqueState | undefined;
  persisted: boolean;
}

export interface ResetRequest {
  factoryDefaults: boolean;
}

export interface ResetResponse {
  state: CirqueState | undefined;
}

export interface Request {
  getState?: GetStateRequest | undefined;
  setState?: SetStateRequest | undefined;
  reset?: ResetRequest | undefined;
}

export interface ErrorResponse {
  message: string;
}

export interface Response {
  error?: ErrorResponse | undefined;
  getState?: GetStateResponse | undefined;
  setState?: SetStateResponse | undefined;
  reset?: ResetResponse | undefined;
}
// ...
export const GetStateRequest: MessageFns<GetStateRequest> = {
  encode(_: GetStateRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetStateRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetStateRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<GetStateRequest>): GetStateRequest {
    return GetStateRequest.fromPartial(base ?? {});
  },
  fromPartial(_: DeepPartial<GetStateRequest>): GetStateRequest {
    const message = createBaseGetStateRequest();
    return message;
  },
};
// ...
export const SetStateRequest: MessageFns<SetStateRequest> = {
  encode(message: SetStateRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.state !== undefined) {
      CirqueState.encode(message.state, writer.uint32(10).fork()).join();
    }
    if (message.persist !== false) {
      writer.uint32(16).bool(message.persist);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SetStateRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSetStateRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.state = CirqueState.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.persist = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<SetStateRequest>): SetStateRequest {
    return SetStateRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<SetStateRequest>): SetStateRequest {
    const message = createBaseSetStateRequest();
    message.state = (object.state !== undefined && object.state !== null)
      ? CirqueState.fromPartial(object.state)
      : undefined;
    message.persist = object.persist ?? false;
    return message;
  },
};
// ...
export const ResetRequest: MessageFns<ResetRequest> = {
  encode(message: ResetRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.factoryDefaults !== false) {
      writer.uint32(8).bool(message.factoryDefaults);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ResetRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseResetRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.factoryDefaults = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<ResetRequest>): ResetRequest {
    return ResetRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ResetRequest>): ResetRequest {
    const message = createBaseResetRequest();
    message.factoryDefaults = object.factoryDefaults ?? false;
    return message;
  },
};
// ...
export const Request: MessageFns<Request> = {
  encode(message: Request, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.getState !== undefined) {
      GetStateRequest.encode(message.getState, writer.uint32(10).fork()).join();
    }
    if (message.setState !== undefined) {
      SetStateRequest.encode(message.setState, writer.uint32(18).fork()).join();
    }
    if (message.reset !== undefined) {
      ResetRequest.encode(message.reset, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Request {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.getState = GetStateRequest.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.setState = SetStateRequest.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.reset = ResetRequest.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<Request>): Request {
    return Request.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Request>): Request {
    const message = createBaseRequest();
    message.getState = (object.getState !== undefined && object.getState !== null)
      ? GetStateRequest.fromPartial(object.getState)
      : undefined;
    message.setState = (object.setState !== undefined && object.setState !== null)
      ? SetStateRequest.fromPartial(object.setState)
      : undefined;
    message.reset = (object.reset !== undefined && object.reset !== null)
      ? ResetRequest.fromPartial(object.reset)
      : undefined;
    return message;
  },
};

function createBaseErrorResponse(): ErrorResponse {
  return { message: "" };
}

export const ErrorResponse: MessageFns<ErrorResponse> = {
  encode(message: ErrorResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.message !== "") {
      writer.uint32(10).string(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ErrorResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseErrorResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.message = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<ErrorResponse>): ErrorResponse {
    return ErrorResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ErrorResponse>): ErrorResponse {
    const message = createBaseErrorResponse();
    message.message = object.message ?? "";
    return message;
  },
};

function createBaseResponse(): Response {
  return { error: undefined, getState: undefined, setState: undefined, reset: undefined };
}

export const Response: MessageFns<Response> = {
  encode(message: Response, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.error !== undefined) {
      ErrorResponse.encode(message.error, writer.uint32(10).fork()).join();
    }
    if (message.getState !== undefined) {
      GetStateResponse.encode(message.getState, writer.uint32(18).fork()).join();
    }
    if (message.setState !== undefined) {
      SetStateResponse.encode(message.setState, writer.uint32(26).fork()).join();
    }
    if (message.reset !== undefined) {
      ResetResponse.encode(message.reset, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Response {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.error = ErrorResponse.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.getState = GetStateResponse.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.setState = SetStateResponse.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.reset = ResetResponse.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<Response>): Response {
    return Response.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Response>): Response {
    const message = createBaseResponse();
    message.error = (object.error !== undefined && object.error !== null)
      ? ErrorResponse.fromPartial(object.error)
      : undefined;
    message.getState = (object.getState !== undefined && object.getState !== null)
      ? GetStateResponse.fromPartial(object.getState)
      : undefined;
    message.setState = (object.setState !== undefined && object.setState !== null)
      ? SetStateResponse.fromPartial(object.setState)
      : undefined;
    message.reset = (object.reset !== undefined && object.reset !== null)
      ? ResetResponse.fromPartial(object.reset)
      : undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}